Big Update
I added the shopping list and the shopping cart list. I gave labels for the lists. I added an "Item not found" label. I reorganized the lists a bit. I added the item info to the search list buttons. Next steps are to make delete functions so that all new searches in the search list delete the old search and so that the user can delete items from the shopping list and the shopping cart. Additionally, the search list only displays the search item title and not the actual item titles.